Fix for msvc compile, lua_bit.cpp wont try to compile if lua isn't set to build

+++ zone/QuestParserCollection.cpp
@@ -196,11 +196,11 @@ bool QuestParserCollection::ItemHasQuestSub(ItemInst *itm, QuestEventID evt) {
 	std::string item_script;
 	if(itm->GetItem()->ScriptFileID != 0) {
 		item_script = "script_";
-		item_script += std::to_string(itm->GetItem()->ScriptFileID);
+		item_script += std::to_string(static_cast<long long>(itm->GetItem()->ScriptFileID));
 	} else if(strlen(itm->GetItem()->CharmFile) > 0) {
 		item_script = itm->GetItem()->CharmFile;
 	} else {
-		item_script = std::to_string(itm->GetID());
+		item_script = std::to_string(static_cast<long long>(itm->GetID()));
 	}
 
 	uint32 item_id = itm->GetID();
@@ -342,11 +342,11 @@ int QuestParserCollection::EventItem(QuestEventID evt, Client *client, ItemInst 
 	std::string item_script;
 	if(item->GetItem()->ScriptFileID != 0) {
 		item_script = "script_";
-		item_script += std::to_string(item->GetItem()->ScriptFileID);
+		item_script += std::to_string(static_cast<long long>(item->GetItem()->ScriptFileID));
 	} else if(strlen(item->GetItem()->CharmFile) > 0) {
 		item_script = item->GetItem()->CharmFile;
 	} else {
-		item_script = std::to_string(item->GetID());
+		item_script = std::to_string(static_cast<long long>(item->GetID()));
 	}
 
 	uint32 item_id = item->GetID();
